created context and provider and custom hook for table data
