+++ src/context/TableDataContext.tsx
@@ -1,5 +1,5 @@
-import { createContext, useState, useContext } from "react";
-import type { ReactNode } from "react";
+import { createContext } from "react";
+
 import type { TableData } from "../types/Types";
 
 // Typ för Context
@@ -9,24 +9,6 @@ type TableDataContextType = {
 };
 
 // Skapa context
-const TableDataContext = createContext<TableDataContextType | undefined>(
+export const TableDataContext = createContext<TableDataContextType | undefined>(
   undefined
 );
-
-// Provider-komponent
-export function TableDataProvider({ children }: { children: ReactNode }) {
-  const [tableData, setTableData] = useState<TableData>([]);
-  return (
-    <TableDataContext.Provider value={{ tableData, setTableData }}>
-      {children}
-    </TableDataContext.Provider>
-  );
-}
-
-// Hook för enklare åtkomst
-export function useTableData() {
-  const context = useContext(TableDataContext);
-  if (!context)
-    throw new Error("useTableData must be used within a TableDataProvider");
-  return context;
-}
