+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.tsx";
-import { TableDataProvider } from "./context/TableDataContext.tsx";
+import { TableDataProvider } from "./context/TableDataProvider.tsx";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
